// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/planProcessStatus.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type PlanProcessStatus = 0 | 1;

export const PlanProcessStatus = {
    NUMBER_0: 0 as PlanProcessStatus,
    NUMBER_1: 1 as PlanProcessStatus
};